python_server/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-multipart==0.0.6
python-dotenv==1.0.0
openai==1.3.7
nbformat==5.9.2
requests==2.31.0
pandas==2.1.0
openpyxl==3.1.2 